+++ lambdas/delete-product/delete-product.js
@@ -1,5 +1,5 @@
 const { getBaseResponse, createEsClient, getId } = require('products-api-utils')
-const deleteEntry = require('./helpers/delete-entry.js')
+const deleteEntry = require('./helpers/delete.js')
 
 module.exports.handler = async (event) => {
   return await deleteEntry(createEsClient(), getId(event), getBaseResponse())
