+++ lambdas/delete-product/helpers/delete.js
