+++ lambdas/get-product/get-product.js
@@ -1,5 +1,5 @@
 const { getBaseResponse, createEsClient, getId } = require('products-api-utils')
-const getEntry = require('./helpers/get-entry.js')
+const getEntry = require('./helpers/get.js')
 
 module.exports.handler = async (event) => {
   return await getEntry(createEsClient(), getId(event), getBaseResponse())
